Huge renaming reshuffle
Okay, I made a few mistakes. When the project started, I tried to put
everything in its own folder. But if you look at any of the big
projects, none of the python files assume that they'll be executed in
the project's root folder.

Most stuff only seems to work after it's been "installed", even
temporary. As such, I've moved all the current "working" files, into a
new folder called sidereal, and gotten rid of the rest of the crud.

In the next few commits, I'll work on getting a basic distutils up, and
a test suite, that'll probably contain a bunch of poorly made unittests.

I also removed a large amount of now unneeded Panda3d stuff, which can
be recovered if we need to get it out. It's the magic of git!