Movement changes to make physics modular
Trying out PyODE as ODE's documentation is better than Bullet, and
getting a working prototype is more important than it being fast. But we
shouldn't chuck out the physics code already written, but see if we can
have the physics modular and interchangeable.

Will probably write an abstract class that everyone has to use, in order
to do the same interface.