Added vaguely spaceship-like Blender model
This is lopsided, and probably a terrible model, but something with
definite shape, and front and back will probably be useful for testing.

My blender skills aren't terribly good however.

I was looking at:
http://s0lar1x.deviantart.com/art/Spaceship-design-73109107
But my modelling skills aren't up to scratch, and it somewhat deviated
from the original picture.