Scrapped ships, modified universe
The ships module is pointless. Since the actual information that
describes an ingame agent (or gasau) is spread over multiple different
objects, having a single point of reference doesn't make sense.

All that the Ship class currently consists of, is a unique id, and
associated logger, which can be obtained from the /logger/ module
anyway.