Log tastic!
The server now prints out DEBUG messages, like every packet that it
sends. The complaints about sequence numbers and unanswered packets are
now also logging methods.

Right now we're logging to the root logger, which isn't the best thing,
but is a quick solution. We'll probably set up a logger hierichy[sic] later.